Move DTOs to schemas
