+++ backend/models/__init__.py
@@ -1,4 +1,2 @@
-from .game import GameDTO, GameCreatedResponse, GameListItemDTO
-from .player import PlayerDTO, ScoresDTO, EndStatsDTO
-from .award import AwardResult
+
 from .enums import MapName, Expansion
