+++ backend/tests/test_records.py
@@ -3,8 +3,7 @@ from tests.fakes import (FakeGamesRepositoryForRecords,
                         make_player)
 from services.records_service import RecordsService
 from schemas.records import RecordDTO
-from models.game import GameDTO
-from models.player import PlayerDTO, ScoresDTO, EndStatsDTO
+
